fix: resolve remaining Kotlin CI compilation errors
- Add .toLong() for OkHttp connectTimeout/readTimeout (Int→Long)
- Make McpBridge.client internal instead of private
- Use safe call ?.online on nullable BridgeState?

+++ app/src/main/java/com/soreverse/mcp/core/McpClient.kt
@@ -58,8 +58,8 @@ class McpClient(
 
     private val client: OkHttpClient by lazy {
         OkHttpClient.Builder()
-            .connectTimeout(connectTimeoutSec, TimeUnit.SECONDS)
-            .readTimeout(readTimeoutSec, TimeUnit.SECONDS)
+            .connectTimeout(connectTimeoutSec.toLong(), TimeUnit.SECONDS)
+            .readTimeout(readTimeoutSec.toLong(), TimeUnit.SECONDS)
             .retryOnConnectionFailure(true)
             .build()
     }
